+++ apps/backend/src/libs/modules/controller/libs/types/api-handler-options.type.ts
@@ -4,6 +4,7 @@ type APIHandlerOptions<
 	body: T["body"];
 	params: T["params"];
 	query: T["query"];
+	uploadedFile?: T["uploadedFile"];
 	user?: T["user"];
 };
 
@@ -11,6 +12,7 @@ type DefaultApiHandlerOptions = {
 	body?: unknown;
 	params?: unknown;
 	query?: unknown;
+	uploadedFile?: unknown;
 	user?: unknown;
 };