+++ apps/backend/src/libs/modules/server-application/libs/types/server-application-route-parameters.type.ts
@@ -14,7 +14,7 @@ type ServerApplicationRouteParameters = {
 	) => Promise<void> | void;
 	method: ValueOf<typeof HTTPMethod>;
 	path: string;
-	preHandlers?: preHandlerHookHandler | preHandlerHookHandler[] | undefined;
+	preHandlers?: preHandlerHookHandler[];
 	validation?: {
 		body?: ValidationSchema;
 		headers?: ValidationSchema;