+++ apps/bot/src/libs/modules/zoom/base-zoom-bot.ts
@@ -184,11 +184,6 @@ class BaseZoomBot {
 				SocketEvent.JOIN_ROOM,
 				String(this.config.ENV.ZOOM.MEETING_ID),
 			);
-			this.logger.info(ZoomBotMessage.GETTING_PUBLIC_URL);
-			this.socketClient.emit(
-				SocketEvent.GET_PUBLIC_URL,
-				String(this.config.ENV.ZOOM.MEETING_ID),
-			);
 		});
 
 		this.socketClient.on(SocketEvent.DISCONNECT, (reason: string) => {
@@ -320,7 +315,11 @@ class BaseZoomBot {
 			this.browser = await puppeteer.launch(this.config.getLaunchOptions());
 			this.page = await this.browser.newPage();
 			await this.page.setUserAgent(USER_AGENT);
-
+			this.initSocket();
+			this.socketClient.emit(
+				SocketEvent.JOINING_TO_MEETING,
+				String(this.config.ENV.ZOOM.MEETING_ID),
+			);
 			this.logger.info(
 				`${ZoomBotMessage.NAVIGATION_TO_ZOOM} ${this.config.ENV.ZOOM.MEETING_LINK}`,
 			);
@@ -334,13 +333,25 @@ class BaseZoomBot {
 			await this.handleInitialPopups();
 			await this.joinMeeting();
 			this.logger.info(ZoomBotMessage.JOINED_MEETING);
-			this.initSocket();
+			this.socketClient.emit(
+				SocketEvent.RECORDING,
+				String(this.config.ENV.ZOOM.MEETING_ID),
+			);
+			this.logger.info(ZoomBotMessage.GETTING_PUBLIC_URL);
+			this.socketClient.emit(
+				SocketEvent.GET_PUBLIC_URL,
+				String(this.config.ENV.ZOOM.MEETING_ID),
+			);
 			this.audioRecorder.start();
 			this.audioRecorder.startFullMeetingRecording(
 				String(this.config.ENV.ZOOM.MEETING_ID),
 			);
 			this.logger.info(ZoomBotMessage.AUDIO_RECORDING_STARTED);
 		} catch (error) {
+			this.socketClient.emit(
+				SocketEvent.FAILED_TO_JOIN_MEETING,
+				String(this.config.ENV.ZOOM.MEETING_ID),
+			);
 			this.logger.error(
 				`${ZoomBotMessage.FAILED_TO_JOIN_MEETING} ${error instanceof Error ? error.message : String(error)}`,
 			);