+++ apps/frontend/src/libs/components/meeting-pdf/meeting-pdf.tsx
@@ -3,9 +3,10 @@ import { DateTimeFormat } from "~/libs/enums/enums.js";
 import { formatDate } from "~/libs/helpers/helpers.js";
 import { type MeetingPdfProperties } from "~/libs/types/types.js";
 
-import { styles } from "./meeting-pdf.styles.js";
+import { renderMarkdown } from "./libs/components/render-markdown.js";
+import { styles } from "./libs/styles/meeting-pdf.styles.js";
 
-const MeetingPdf = ({
+const MeetingPdf: React.FC<MeetingPdfProperties> = ({
 	actionItems,
 	createdAt,
 	summary,
@@ -14,30 +15,24 @@ const MeetingPdf = ({
 }: MeetingPdfProperties): React.ReactElement => {
 	return (
 		<Document>
-			<Page size="A4" style={styles.page}>
-				<View style={styles.section}>
-					<Text style={styles.title}>{title}</Text>
-					<Text style={styles.text}>
-						Date: {formatDate(new Date(createdAt), DateTimeFormat.MEETING)}
-					</Text>
-				</View>
-				<View style={styles.section}>
-					<Text style={styles.title}>Summary</Text>
-					<Text style={styles.text}>{summary}</Text>
-				</View>
+		<Page size="A4" style={styles.page}>
+			<View style={styles.section}>
+				<Text style={styles.title}>{title}</Text>
+				<Text style={styles.text}>
+					Date: {formatDate(new Date(createdAt), DateTimeFormat.MEETING)}
+				</Text>
+			</View>
 
-				<View style={styles.section}>
-					<Text style={styles.title}>Action Items</Text>
-					<Text style={styles.text}>{actionItems}</Text>
-				</View>
+			<View style={styles.section}>{renderMarkdown(summary)}</View>
 
-				<View style={styles.section}>
-					<Text style={styles.title}>Transcript</Text>
-					<Text style={styles.text}>{transcription}</Text>
-				</View>
-			</Page>
-		</Document>
-	);
-};
+			<View style={styles.section}>{renderMarkdown(actionItems)}</View>
+
+			<View style={styles.section}>
+				<Text style={styles.heading1}>Transcript</Text>
+				{renderMarkdown(transcription)}
+			</View>
+		</Page>
+	</Document>
+);
 
 export { MeetingPdf };