+++ packages/shared/src/index.ts
@@ -1,6 +1,5 @@
 export {
 	EMPTY_ARRAY_LENGTH,
-	MILLISECONDS_IN_SECOND,
 	PERCENT_MULTIPLIER,
 	START_TIME,
 } from "./libs/constants/constants.js";
@@ -21,13 +20,17 @@ export {
 	HTTPError,
 	MeetingError,
 	MeetingTranscriptionError,
+	S3Error,
+	UploadError,
 	UserError,
 	ValidationError,
 } from "./libs/exceptions/exceptions.js";
 export {
 	configureString,
+	convertToZoomWebClientUrl,
 	extractZoomMeetingId,
 	formatDate,
+	isZoomLinkValid,
 } from "./libs/helpers/helpers.js";
 export { type Config } from "./libs/modules/config/config.js";
 export {
@@ -52,6 +55,7 @@ export {
 	type ValueOf,
 } from "./libs/types/types.js";
 export { AuthApiPath, AuthStatusMessage } from "./modules/auth/auth.js";
+export { S3ErrorMessage } from "./modules/aws/aws.js";
 export {
 	type FileGetAllResponseDto,
 	type FilePublicDto,
@@ -59,6 +63,7 @@ export {
 	type FileResponseDto,
 	type FileUpdateRequestDto,
 	FileErrorMessage,
+	UploadErrorMessage,
 } from "./modules/files/files.js";
 export {
 	type MeetingAttachAudioRequestDto,
@@ -69,6 +74,7 @@ export {
 	type MeetingGetAllResponseDto,
 	type MeetingGetPublicUrlResponseDto,
 	type MeetingResponseDto,
+	type MeetingStatusDto,
 	type MeetingSummaryActionItemsResponseDto,
 	type MeetingTranscriptionGetAllResponseDto,
 	type MeetingTranscriptionRequestDto,
@@ -87,6 +93,9 @@ export {
 } from "./modules/meetings/meetings.js";
 export {
 	type AuthResponseDto,
+	type AvatarFileDto,
+	type AvatarUploadResponseDto,
+	type DeleteAvatarResult,
 	type UserGetAllItemResponseDto,
 	type UserGetAllResponseDto,
 	type UserResponseDto,