+++ packages/shared/src/libs/enums/exception-message.enum.ts
@@ -6,6 +6,8 @@ const ExceptionMessage = {
 	FORBIDDEN: "Access denied",
 	MEETING_FAILED_TO_CREATE: "Failed to create a meeting",
 	MEETING_NOT_FOUND: "Meeting not found",
+	METHOD_NOT_ALLOWED: "Method not allowed for this route",
+	ROUTE_NOT_FOUND: "Route not found",
 	UPDATE_FAILED: "Failed to update meeting",
 } as const;