+++ packages/shared/src/modules/meetings/libs/enums/meeting-error-message.enum.ts
@@ -13,8 +13,11 @@ const MeetingErrorMessage = {
 	MEETING_FAILED_TO_CREATE: "Failed to create a meeting",
 	MEETING_NOT_FOUND: "Meeting not found",
 	MEETING_SUMMARY_NOT_AVAILABLE: "Meeting summary not available.",
+	NO_MEETING_FOR_LINK: "There is no active meeting for this link.",
 	SEARCH_NOT_EMPTY: "Search input cannot be empty.",
 	UPDATE_FAILED: "Failed to update meeting",
+	USER_IS_REQUIRED_TO_CREATE_MEETING: "User is required to create a meeting",
+	USER_IS_REQUIRED_TO_FIND_MEETINGS: "User is required to create a meeting",
 } as const;
 
 export { MeetingErrorMessage };